remove ParamLookup from Printer
Remove the ability to customize the serialization of arbitrary bound variables and just serialize them by appending the nonce to the end of the variable name. Also for let-bound variables use a name that is an invalid crucible variable name, so we're guaranteed to get no collissions (specifically `@l` followed by the nonce id).